+++ webrtc/src/webrtc_c/rtc_base/crypto_random.cc
@@ -14,7 +14,8 @@
 // -------------------------
 
 extern "C" {
-WEBRTC_EXPORT struct std_string_unique* webrtc_CreateRandomString(int length) {
+WEBRTC_EXPORT struct std_string_unique* webrtc_CreateRandomString(
+    size_t length) {
   auto str = std::make_unique<std::string>(webrtc::CreateRandomString(length));
   return reinterpret_cast<struct std_string_unique*>(str.release());
 }